+++ CHANGELOG.rst
@@ -8,6 +8,11 @@ Changes
 -------
 Unreleased
 ==========
+2019-10-30: v0.5.8
+^^^^^^^^^^^^^^^^^^
+* :hammer: Prefix all [reprolib](https://github.com/ReproNim/schema-standardization) terms
+* :ambulance: :tractor: Delanguage tag `http://schema.org/url`s
+
 2019-10-25: v0.5.7
 ^^^^^^^^^^^^^^^^^^
 * :racehorse: Improve `GET user/applets` speed